pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openapply"
version = "0.1.0"
description = "Open-source autonomous job search agent. Runs locally."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "rich>=13",
  "textual>=0.80",
  "ollama>=0.3",
  "playwright>=1.44",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "pyyaml>=6",
  "jinja2>=3.1",
  "httpx>=0.27",
  "python-slugify",
  "python-dateutil",
]

[project.scripts]
openapply = "cli.main:app"

[tool.setuptools.package-data]
agent = ["prompts/*.md", "templates/*.html"]
cli = ["assets/*.yml"]

[tool.setuptools.packages.find]
include = ["agent", "agent.*", "cli", "cli.*", "memory", "memory.*"]
